refactor: split monolithic sla_breach_detector into modular components
  * Extracted business logic into services.py
  * Moved API routes to routes.py
  * Isolated Prometheus metrics in metrics.py
